Split `insert` into sub-modules
